// src/app/domain/model/domain/education-domain.model.ts
import { GradeEntityDomainType } from '../../type/entity/grade-entity-domain.type';
import { EducationKindDomainType } from '../../type/kind/education-kind-domain.type';
import { DegreeNameDomainType } from '../../type/name/degree-name-domain.type';
import { FieldOfStudyNameDomainType } from '../../type/name/field-of-study-name-domain.type';
import { SchoolNameDomainType } from '../../type/name/school-name-domain.type';
import { BaseDomain } from '../../../data/model/base/base-data.model';

export namespace EducationDomain {
  export interface Data extends BaseDomain.Data {
    school: SchoolNameDomainType;
    degree: DegreeNameDomainType;
    fieldOfStudy: FieldOfStudyNameDomainType;
    startYear: number;
    endYear: number | null;
    grade: GradeEntityDomainType;
  }

  export interface Metadata extends BaseDomain.Metadata {
    kind: EducationKindDomainType;
    order: number;
  }
}